Register volume button on player as an obstruction, add logic for volume control (mute/unmute ad) and update insideAdVolumeChanged event

+++ app/src/main/java/com/streann/insidead/demo/MainActivity.kt
@@ -51,7 +51,7 @@ class MainActivity : AppCompatActivity() {
                     adProgressText.text = "Error: $error \nShow Ad"
                 }
 
-                override fun insideAdVolumeChanged(level: Float) {
+                override fun insideAdVolumeChanged(level: Int) {
                     Log.i(TAG, "insideAdVolumeChanged: $level")
                 }
             })
